% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean's Algorithmm}
\usage{
euclidean(input1, input2)
}
\arguments{
\item{input1}{a numeric scalar}

\item{input2}{a numeric scalar}
}
\value{
The function returns the greatest common divisor of the given two inputs
}
\description{
This function calculates the greatest common divisor, that is the
number that divides both inputs without remainder
}
\examples{
euclidean(123612, 13892347912)
}
\seealso{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Euclidean Algorithm}
}
