% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia graph}
\format{
wiki_graph
A data frame with 3 columns:
\describe{
  \item{v1}{from node}
  \item{v2}{to node}
  \item{w}{weight of the connection}
}
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm/
}
\usage{
wiki_graph
}
\description{
A dataset containing a connected graph from the wikipedia page of Dijkstra'a algorithm meant to test the functionning of the dijkstra function in this package.
}
\keyword{datasets}
